Added test_hybrid_set.cc, test_discretiser.cc and test_reachability_analyser.cc. Changed name of GridCell::smallest_primary_cell(...) to smallest_primary_cell_height and removed dimensions argument since this can be deduced from the box. Added GridCell over_approximation(Box,Grid) method. Added missing code to Orbit classes. Added some missing code to Discretiser class.

+++ src/matrix.cc
@@ -33,8 +33,8 @@ namespace Ariadne {
 Matrix<Float> 
 midpoint(const Matrix<Interval>& A) {
   Matrix<Float> R(A.row_size(),A.column_size());
-  for(uint i=0; i!=A.row_size(); ++i) {
-    for(uint j=0; j!=A.row_size(); ++j) {
+  for(size_t i=0; i!=A.row_size(); ++i) {
+    for(size_t j=0; j!=A.row_size(); ++j) {
       R[i][j]=A[i][j].midpoint();
     }
   }
