+++ src/vector.cc
@@ -33,7 +33,7 @@ namespace Ariadne {
 bool subset(const Vector<Float>& v1, const Vector<Interval>& v2)
 {
   ARIADNE_ASSERT(v1.size()==v2.size());
-  for(uint i=0; i!=v1.size(); ++i) {
+  for(size_t i=0; i!=v1.size(); ++i) {
     if(!subset(v1[i],v2[i])) { return false; }
   }
   return true;
@@ -42,7 +42,7 @@ bool subset(const Vector<Float>& v1, const Vector<Interval>& v2)
 bool subset(const Vector<Interval>& v1, const Vector<Interval>& v2) 
 {
   ARIADNE_ASSERT(v1.size()==v2.size());
-  for(uint i=0; i!=v1.size(); ++i) {
+  for(size_t i=0; i!=v1.size(); ++i) {
     if(!subset(v1[i],v2[i])) { return false; }
   }
   return true;
@@ -51,7 +51,7 @@ bool subset(const Vector<Interval>& v1, const Vector<Interval>& v2)
 Vector<Float> midpoint(const Vector<Interval>& v) 
 {
   Vector<Float> r(v.size());
-  for(uint i=0; i!=v.size(); ++i) {
+  for(size_t i=0; i!=v.size(); ++i) {
     r[i]=v[i].midpoint();
   }
   return r;
@@ -60,7 +60,7 @@ Vector<Float> midpoint(const Vector<Interval>& v)
 Vector<Float> lower(const Vector<Interval>& v) 
 {
   Vector<Float> r(v.size());
-  for(uint i=0; i!=v.size(); ++i) {
+  for(size_t i=0; i!=v.size(); ++i) {
     r[i]=v[i].lower();
   }
   return r;
@@ -69,7 +69,7 @@ Vector<Float> lower(const Vector<Interval>& v)
 Vector<Float> upper(const Vector<Interval>& v) 
 {
   Vector<Float> r(v.size());
-  for(uint i=0; i!=v.size(); ++i) {
+  for(size_t i=0; i!=v.size(); ++i) {
     r[i]=v[i].upper();
   }
   return r;
@@ -79,7 +79,7 @@ Vector<Interval> hull(const Vector<Interval>& v1, const Vector<Interval>& v2)
 {
   ARIADNE_ASSERT(v1.size()==v2.size());
   Vector<Interval> r(v1.size());
-  for(uint i=0; i!=v1.size(); ++i) {
+  for(size_t i=0; i!=v1.size(); ++i) {
     r[i]=hull(v1[i],v2[i]);
   }
   return r;
@@ -89,7 +89,7 @@ Vector<Interval> intersection(const Vector<Interval>& v1, const Vector<Interval>
 {
   ARIADNE_ASSERT(v1.size()==v2.size());
   Vector<Interval> r(v1.size());
-  for(uint i=0; i!=v1.size(); ++i) {
+  for(size_t i=0; i!=v1.size(); ++i) {
     r[i]=intersection(v1[i],v2[i]);
   }
   return r;
@@ -98,7 +98,7 @@ Vector<Interval> intersection(const Vector<Interval>& v1, const Vector<Interval>
 Float radius(const Vector<Interval>& v) 
 {
   Float r=0;
-  for(uint i=0; i!=v.size(); ++i) {
+  for(size_t i=0; i!=v.size(); ++i) {
     r=Ariadne::max(r,v[i].radius());
   }
   return r;
@@ -107,7 +107,7 @@ Float radius(const Vector<Interval>& v)
 Float volume(const Vector<Interval>& v) 
 {
   Float r=1.0;
-  for(uint i=0; i!=v.size(); ++i) {
+  for(size_t i=0; i!=v.size(); ++i) {
     r*=diam(v[i]);
   }
   return r;
@@ -116,7 +116,7 @@ Float volume(const Vector<Interval>& v)
 bool disjoint(const Vector<Interval>& v1, const Vector<Interval>& v2) 
 {
   ARIADNE_ASSERT(v1.size()==v2.size());
-  for(uint i=0; i!=v1.size(); ++i) {
+  for(size_t i=0; i!=v1.size(); ++i) {
     if(v1[i].u<v2[i].l || v1[i].l>v2[i].u) {
       return true;
     }
